requirements.txt
fastapi
motor
pydantic
python-dotenv
requests
dnspython
python-whois
beautifulsoup4
lxml
aiohttp
cachetools
uvicorn
jwt
PyJWT